requirements.txt
gradio>=4.36.0
google-genai>=1.0.0
anthropic>=0.40.0
openai>=1.50.0
